rename document-browser to virtual-dom
